Print version on startup for firejail/firecfg
It is not too uncommon for the firejail version to be missing when
issues are reported; this commit makes it more likely that any posted
logs will contain the program version.

Do so just for firejail and firecfg for now because they are the most
common user-facing programs.

Print the version after argument parsing, in order to avoid printing the
program version more than once and to avoid interfering with commands
that generate machine-readable output (like `firejail --list` and
`firecfg --list`).  Also, only print it after all profiles have been
loaded, because a profile may contain `quiet`.

Note: This does not cover the case where the program exits before the
end of argument/profile parsing (such as when an error occurs).