ci: move main code checks into new check-c.yml
Move scan-build, cppcheck and CodeQL (cpp).

This is similar to build-extra.yml, but for jobs that check for issues
in the code rather than checking for build failures.

Note: As this deletes codeql-analysis.yml, its configuration also has to
be deleted in the GitHub web UI to prevent it from warning about the
file being missing:

* Security -> Code scanning -> Tool status -> (Setup Types) CodeQL ->
  (Configurations) language:python -> Delete configuration

Misc: The above was clarified by @topimiettinen[1].

[1] https://github.com/netblue30/firejail/pull/5960#issuecomment-1685262643