tests: properly fix fs/kmsg test
It was broken likely due to `private-dev` being added to default.profile
on commit 307dad542 ("adding private-tmp and private-dev to
default.profile", 2023-08-20).

So ignore `private-dev` in the test and make sure to run the tests when
default.profile changes.

This amends commit 75cefd5b1 ("tests: fix error when /dev/kmsg is
missing", 2023-08-21).