ci: fix dependabot duplicated workflow runs
Every workflow is being executed twice for dependabot: Once when its
branch is pushed to this repository and again when a PR is opened for
it.

For example, see the checks in #5979 ("29 checks passed").

This happens because both `on.push` and `on.pull_request` are specified
in the workflow files.

There does not seem to be a simple and generic way to avoid such
duplicated runs directly in GitHub Actions (such as preventing the same
check from running for the same exact commit)[1], so just ignore the
dependabot branches on push for now.

See also and commit 5871b08a4 ("ci: run for every branch instead of just
master", 2023-04-23) / PR #5815.

[1] https://github.com/orgs/community/discussions/26276