ci: re-enable sort.py
It was disabled on commit df6ea884f ("merges, disable sort.py in profile
checks temporarely, two more private-etc profiles", 2023-02-14).

Currently all profiles are sorted and there are no ongoing `private-etc`
changes, so it should be safe to re-enable.

Note that the script is useful to catch sorting issues not only in
`private-etc` but also in other commands, such as `seccomp`[1] [2].

This is a follow-up to #6070.

Relates to #5610.

[1] https://github.com/netblue30/firejail/pull/6066#discussion_r1372055800
[2] https://github.com/netblue30/firejail/pull/6067#discussion_r1372027243