profiles: use only /usr/share/lua*
To ensure that it includes luajit paths as well:

* /usr/share/lua
* /usr/share/luajit-2.1

And remove all entries of the same path without the wildcard, to avoid
redundancy.

Misc: The wildcard entries were added on commit 56b60dfd0 ("additional
Lua blacklisting (#3246)", 2020-02-24) and the entries without the
wildcard were partially removed on commit 721a984a5 ("Fix Lua in
disable-interpreters.inc", 2020-02-24).

This is a follow-up to #6128.

Reported-by: @pirate486743186