Fix duplicate address detection
The current code compares the target IP of an ARP reply packet with the
srcaddr/sender IP of the request packet. Instead it should compare the
source IP of the ARP reply with the destaddr/target IP of the request
packet. In addition, this adds the check for ARP requests with the same
IP address in case other systems are probing for the same IP at the same
time as per the recommendation in RFC 5227.

This is related to #6133.