feature: use globbing in hardcoded numbered /dev paths
Certain devices appear in numbered paths in /dev (such as `/dev/hidraw0`
or `/dev/hidraw1`) when they are plugged in.  When private-dev is used,
it attempts to include a hardcoded list of /dev paths in the sandbox.

Since the hardcoded paths only go up to a specific number (mostly from 0
to 9), devices that appear at higher numbered paths (such as
`/dev/hidraw10` or `/dev/hidraw20`) do not show up in /dev when
private-dev is enabled.

This issue also affects options that attempt to disable numbered paths
(such as --no3d and --nou2f), independently of whether private-dev is
used.

Fix this by using glob patterns (such as `/dev/hidraw[0-9]*`) to ensure
that all relevant numbered paths are matched regardless of number.

The globbing is similar to the one from commit 2993298aa ("firecfg:
parse config files in /etc/firejail/firecfg.d", 2023-06-29) / PR #5876.

Closes #2723.

Suggested-by: @tsankuanglee