Revert "build: move HAVE_SANDBOX_CHECK out of MANFLAGS"
This reverts commit 5c6fa6ab584eb6241467b5d5aea76f26dcc922fb.

The commit in question causes `HAVE_SANDBOX_CHECK` to always be unset
(instead of only when `--disable-sandbox-check` is used), as its value
was being passed to the compiler through `MANFLAGS`.  Move the macro
back into `MANFLAGS` for simplicity.

Also, using `--disable-sandbox-check` breaks the tests and thus also
breaks CI (see #6619).

Relates to #6592.