ci: drop apt dependencies from main build
Disable apparmor/selinux and use the default compiler version (currently
gcc 11.4.0-1ubuntu1~22.04).

Note that these configuration options are are still enabled in the
`build-gcc` job, so the code coverage should remain unaffected.

This allows dropping the apt-get update/install commands, which can take
a significant amount of time (sometimes even most of the job runtime),
depending on the state of the CI infrastructure.

For a relatively recent example of this, see the build of commit
a31f74165 ("bugfix: fix "Not enforcing Landlock" message always being
printed (#6806)", 2025-07-10) [1]:

* `apt-get update`: 12s
* `apt-get install`: 64s
* `make`: 40s

When comparing master (8bccd686f) [2] [3] [4] to this commit [5] [6] [7]
using 3 consecutive builds each, the average build time goes down from
77s to 21s (a ~72% reduction), making the CI feedback much quicker,
which is especially helpful when iterating on a branch in general and
when trying to debug CI build failures.

[1] https://github.com/netblue30/firejail/actions/runs/16141617993/job/45550207173
[2] https://github.com/netblue30/firejail/actions/runs/17019049268/job/48245518988
[3] https://github.com/netblue30/firejail/actions/runs/17019049268/job/48245547046
[4] https://github.com/netblue30/firejail/actions/runs/17019049268/job/48245579009
[5] https://github.com/netblue30/firejail/actions/runs/17019030502/job/48245478097
[6] https://github.com/netblue30/firejail/actions/runs/17019030502/job/48245617867
[7] https://github.com/netblue30/firejail/actions/runs/17019030502/job/48245626724