profiles: seahorse: add redirect org.gnome.seahorse.Application
Apparently the .desktop file for `seahorse` is located in the following
path:

* `/usr/share/applications/org.gnome.seahorse.Application.desktop`

Which ends in `Application.desktop` instead of `seahorse.desktop`,
leading to it not being automatically detected by firecfg.

So add a redirect profile and an entry in firecfg.config.

See also commit 8f69e9841 ("bugfix: firecfg: check full filename in
check_profile() (#6674)", 2025-03-04).

Fixes #6658.

Reported-by: @ginto37
Reported-by: @rusty-snake