profiles: add more xorg paths
Add the following files, which may be used to configure X clients:

* `~/.Xdefaults`
* `~/.Xdefaults-*` (`~/.Xdefaults-$(hostname)`)
* `~/.Xresources`

And block the following paths, which are intended for the X server:

* `~/.local/share/xorg` (rootless Xorg log directory)
* `/etc/X11/xinit`
* `/etc/X11/xorg.conf.d`
* `/var/log/Xorg.*` (default Xorg log path)

Note: ~/.Xdefaults is read directly by each application when it starts,
while ~/.Xresources is loaded once into the X root window with xrdb(1)
when starting the session, such as by a DE or directly in ~/.xinitrc.
Both use the same format and it appears that users are encouraged to use
~/.Xresources instead of ~/.Xdefaults but applications still try to read
~/.Xdefaults if it exists.

From xrdb(1):

> FILES
>        Xrdb does not load any files on its own, but many desktop
>        environments use xrdb to load ~/.Xresources files on session
>        startup to initialize the resource database, as a generalized
>        replacement for ~/.Xdefaults files.

See X(1), Xorg(1), xinit(1) and xrdb(1).